+++ src/main/java/dev/orne/config/impl/EnvironmentConfigImpl.java
@@ -1,0 +1,110 @@
+package dev.orne.config.impl;
+
+/*-
+ * #%L
+ * Orne Config
+ * %%
+ * Copyright (C) 2019 - 2025 Orne Developments
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-3.0.html>.
+ * #L%
+ */
+
+import java.util.Map;
+import java.util.stream.Stream;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+import org.apiguardian.api.API;
+
+import dev.orne.config.Config;
+import dev.orne.config.ConfigException;
+
+/**
+ * Implementation of {@code Config} based on the environment variables.
+ * 
+ * @author <a href="https://github.com/ihernaez">(w) Iker Hernaez</a>
+ * @version 1.0, 2025-07
+ * @since 0.1
+ * @see Config
+ * @see System#getenv()
+ */
+@API(status = API.Status.INTERNAL, since = "1.0")
+public class EnvironmentConfigImpl
+extends AbstractConfig {
+
+    /**
+     * Creates a new instance.
+     * 
+     * @param options The configuration builder options.
+     */
+    public EnvironmentConfigImpl(
+            final @NotNull ConfigOptions options) {
+        super(options);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected boolean isEmptyInt() {
+        return getEnvironmentVariables().isEmpty();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected boolean containsInt(
+            final @NotBlank String key) {
+        return getEnvironmentVariables().containsKey(key);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected @NotNull Stream<String> getKeysInt() {
+        return getEnvironmentVariables().keySet()
+                .stream();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected String getInt(
+            final @NotBlank String key) {
+        return getEnvironmentVariables().get(key);
+    }
+
+    /**
+     * Returns the environment variables.
+     * 
+     * @return The environment variables.
+     * @throws ConfigException If a security manager exists and its
+     * {@code checkPermission} method doesn't allow access to the
+     * process environment.
+     * @see System#getenv()
+     */
+    protected @NotNull Map<String, String> getEnvironmentVariables() {
+        try {
+            return System.getenv();
+        } catch (final SecurityException e) {
+            throw new ConfigException("Cannot access environment variables", e);
+        }
+    }
+}